cdap-2670 move copied file to cdap-etl
